requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
python-multipart
python-docx
sqlmodel
